Add a new test fixture for function-per-test
This fixture duplicates (copypasta-style) the tests from
FunctionPerFixtureFacts, but gives the function spun up a very different
lifetime: a new instance per test. To reduce the instance of integration
issues in this scenario*, the port number is now randomly-generated from
the range 50000..60000 for each test.

\* Xunit.net runs tests in parallel by default, the function could be
spun up with the same port, or torn down mid-test,