Add tests for configured function per fixture
Define a new class fixture `ConfiguredAzureFunctionFixture` to set up
the required configuration before running the tests. There's not a nicer
way in Xunit AFAICT to use a class fixture for a shared function and set
up the configuration outside of that fixture, but this is probably
forcing me down a better path wrt Single Responsibility, etc.

Again, the tests methods themselves are duplicated copypasta-style. It
might be worth at this point deduplicating the three copies into one.