Extract Corvus.Testing.AzureFunctions library (#83)
* Extract new Corvus.AzureFunctions library

This begins the process, moving the FunctionConfiguration class and
copying the appropriate pieces from the Corvus.SpecFlow.Extensions
project file to the new library's project file.

* Move process-wrangling classes to Functions lib

This is a straight R#-enabled move of the classes to the new project,
with namespaces updated along the way.

* Fix R# niggles

Small tweaks, seemingly independent of the house style, to tidy up
warnings before migrating the FunctionsController.

* Begin extraction of FunctionsController

Redefine the class to be partial so that the class can be split whilst
preserving current behaviour (and green tests). Move the private fields
to the emptier partial, and define a new StartFunctionsInstance() method
that is independent of test frameworks.

* Move private methods to new partial

* Call new StartFunctionsInstance() from old

Delegate to the new implementation, which is currently a no-op, so that
the code can be moved over piece-by-piece.

The FunctionConfiguration resolution now needs to happen at the
beginning of the old implementation, so that it can be passed to the new
implementation.

* Extract NPM resolution to helper method

This is a relatively self-contained piece of logic to extract.
Extracting the whole lot in one go caused the demo tests to fail, so I'm
scaling it back to a bit at a time.

* Extract GetWorkingDirectory() function

This should be fairly well-isolated at this point - just need to replace
the TestContext references with a parameter.

* Extract parameter for current directory

* Extract method to start the Function Host process

* Move GetWorkingDirectory() to new implementation

* Move StartFunctionHostProcess() to new impl

* Remove remaining implementation code to new

With the various pieces now in place in the new SpecFlow-independent
FunctionsController, those pieces can all be called from that new
implementation.

* Move remaining public methods to new impl

The TeardownFunctions and GetFunctionsOutput methods now more naturally
belong on the SpecFlow-independent part.

* Add System.Management to AzureFunctions library

This is required for tearing down the function host processes after
their test(s).

* Rename Corvus.AzureFunctions library

Per guidance from @jongeorge1 [via Twitter][1], the preferred name for
this library is Corvus.Testing.AzureFunctions.

  [1]: https://twitter.com/jon_george1/status/1255180531244941313

* Fix up namespaces post-rename

* Replace NUnit guard assertions with exceptions

* Move generalised FunctionsController to new lib

A little bit of jiggery-pokery is required here. First, the two parts of
the FunctionsController class have the `partial` modifier removed;
second, the public methods GetFunctionsOutput() and TeardownFunctions()
are reimplemented as delegating members on the SpecFlow-specific
version; finally, there's a little bit of type-system hackery in
DemoFunctionPerFeatureHooks to convince the compiler to import the
appropriate version of FunctionController. This last can be easily
resolved by renaming one or other of the classes as preferred.

* Extract Start() method for process

A general rule of "constructors should do no work" applies, so extract a
method to separate process configuration from startup.

* Move FunctionConfiguration out of Internal ns

Fix from review.

* Resolve netcoreapp3.1 runtime by default

This is the newest LTS release of .NET Core, so makes sense to be the
default now.

* Implement all Exception constructors

Resolves [RCS1194][1].

  [1]: https://github.com/JosefPihrt/Roslynator/blob/master/docs/analyzers/RCS1194.md

* Add first Xunit demo test

There's a bit of boilerplate here, much as there is for the SpecFlow
demo tests. The difference is this is all in-project as there's no
Corvus.Xunit.Extensions library :upside_down_face:

* Add test for failing GET request

* Add tests for POSTS with query string

* Add first test for POST requests with body

This is the happy path, and includes the boilerplate required to convert
the anonymous type to JSON.

* Complete POST tests and fixture

The remaining two tests Just Work :tada:

* Extract Uri property to reduce duplication

* Add a new test fixture for function-per-test

This fixture duplicates (copypasta-style) the tests from
FunctionPerFixtureFacts, but gives the function spun up a very different
lifetime: a new instance per test. To reduce the instance of integration
issues in this scenario*, the port number is now randomly-generated from
the range 50000..60000 for each test.

\* Xunit.net runs tests in parallel by default, the function could be
spun up with the same port, or torn down mid-test,

* Add tests for configured function per fixture

Define a new class fixture `ConfiguredAzureFunctionFixture` to set up
the required configuration before running the tests. There's not a nicer
way in Xunit AFAICT to use a class fixture for a shared function and set
up the configuration outside of that fixture, but this is probably
forcing me down a better path wrt Single Responsibility, etc.

Again, the tests methods themselves are duplicated copypasta-style. It
might be worth at this point deduplicating the three copies into one.

* Update Corvus.Testing.AzureFunctions.csproj

Added NoWarn to project file to suppress exception constructors warning.

* Removed legacy FunctionsController and added new method to retrieve the new FunctionsController for SpecFlow tests.

* Fix comment typo

* Replace process exception with Function version

Delete the original FunctionStartupException and rename
ProcessStartupException to FunctionStartupException.

* Improve documentation on FunctionStartupException

* Make runtime argument required

It was a breaking change to switch from netcoreapp2.1 to netcoreapp3.1
anyway, so we might as well use this opportunity to make this a required
argument.

* Rename parameter

Co-authored-by: Jonathan George <Jonathan.George@endjin.com>