Allow existing function instances to be used test runs (#99)
* Reuse an existing Function instance for tests

If there's already a process listening on the port requested for the
tests, then reuse that process for the test run. This is useful for
debugging failures in the function itself, where the tests can be run
from the command-line (remembering to include the --no-build flag)
against a Function instance running in Visual Studio with the debugger
attached.

We print a warning indicating the test may produce unexpected result
(e.g. if the process dies mid-run, is a build from a different branch,
etc.). With the function process no longer under the control of the test
process, it's much more likely the test will fail for spurious reasons.

* Extract GetWorkingDirectory to a public method

This is a useful utility method for resolving the Function project
directory.

* Code tidy: push method call down the stack