change function host started logic
Updated the logic used to determine when the function
host is ready to receive http requests. On linux the
previous logic was reporting to early and the http requests
were failing.
The new logic waits until the function host reports the
endpoints of the function so we know that they are ready.