add cross platform support (#272)
* make directory paths case sensitive

This is need to support linux because Linux OS
file systems are case sensitive

* change npm command

Changed the npm command used to find the
azure-function-core-tools bin folder.
Using the 'npm prefix' command reports the wrong folder
on linux. Using the 'npm root' command fixes this.

* change function host started logic

Updated the logic used to determine when the function
host is ready to receive http requests. On linux the
previous logic was reporting to early and the http requests
were failing.
The new logic waits until the function host reports the
endpoints of the function so we know that they are ready.