Handle casing variations on *nix
Although the PR fixed the most common way in which the code breaks on case-sensitive filesystems, it would not work correctly in cases where the build folder contains "debug" (instead of "Debug"), which we do occasionally see.

The PR fixed the original problem (which was that we depended on case insensitivity, and thus broke on Linux and macOS) by removing the code that just forced the entire path to lowercase. But this no longer handles non-standard build config case names. So this commit modifies it to use string-insensitive string comparisons to detect the build config, and then to preserve whatever casing happens to be present on the filesystem.

By working with whatever case happens to be present, we restore the original behaviour (in which this code works for any casing - debug, Debug, or even dEBuG) while also fixing the problem the PR fixed: it should now work correctly on case-sensitive filesystems.

Also made some minor fixes:

 * fixed a typo
 * moved some comments around to prevent StyleCop warnings
 * reworded some comments to be more consistent with house style
 * in ResolvePath (which we're changing anyway in this commit) used a compound ??= assignment to stop VS kvetching