Set IsTestProject to false on non-test projects
The SDK seems to assume that if the project name includes the word "Test" that it should inspect it for tests. Given the name of this library, *every* project meets that criterion, so we get warnings about projects not having test adapters.

So we set the property explicitly to false on projects that don't actually contain any tests. (Most of the projects are libraries intended to support projects that do contain tests.)

This doesn't seem to fix the "Microsoft.VisualStudio.TestPlatform.ObjectModel.TestPlatformException: Could not find testhost" errors, but at least it slightly reduces the chatter in the VS Test output.